LockType.h: rename to LockMode.h
