+++ xbmc/LockMode.h
